paletted image cpp extract
