+++ src/paletted_image.cpp
@@ -1,0 +1,60 @@
+#include <string>
+#include <vector>
+#include <algorithm>
+#include <loguru.hpp>
+#include <stb_image.h>
+#include <stb_image_write.h>
+#include "paletted_image.hpp"
+#include "rgba.hpp"
+#include "constants.hpp"
+
+
+PalettedImage load_paletted_image(const std::string& path)
+{
+	ERROR_CONTEXT("loading sample image", path.c_str());
+	int width, height, comp;
+	RGBA* rgba = reinterpret_cast<RGBA*>(stbi_load(path.c_str(), &width, &height, &comp, 4));
+	CHECK_NOTNULL_F(rgba);
+	const auto num_pixels = width * height;
+
+	// Fix issues with stbi_load:
+	if (comp == 1) {
+		// input was greyscale - set alpha:
+		for (auto& pixel : emilib::it_range(rgba, rgba + num_pixels)) {
+			pixel.a = pixel.r;
+		}
+	} else {
+		if (comp == 3) {
+			for (auto& pixel : emilib::it_range(rgba, rgba + num_pixels)) {
+				pixel.a = 255;
+			}
+		}
+		for (auto& pixel : emilib::it_range(rgba, rgba + num_pixels)) {
+			if (pixel.a == 0) {
+				pixel = RGBA{0,0,0,0};
+			}
+		}
+	}
+
+	std::vector<RGBA> palette;
+	std::vector<ColorIndex> data;
+
+	for (const auto pixel_idx : irange(num_pixels)) {
+		const RGBA color = rgba[pixel_idx];
+		const auto color_idx = std::find(palette.begin(), palette.end(), color) - palette.begin();
+		if (color_idx == palette.size()) {
+			CHECK_LT_F(palette.size(), MAX_COLORS, "Too many colors in image");
+			palette.push_back(color);
+		}
+		data.push_back(color_idx);
+	}
+
+	stbi_image_free(rgba);
+
+	return PalettedImage{
+		static_cast<size_t>(width),
+		static_cast<size_t>(height),
+		data, palette
+	};
+}
+
